test: merge catch_optimization.rs into exceptions.rs
Move 6 unique tests (4-arm primitive types, float wildcard, 3-arm bool,
4 user classes, 4 literals, mixed named/anonymous bindings) into
exceptions.rs with their OptLevel::Two bytecode snapshots preserved.
Delete catch_optimization.rs — its remaining 2 tests were already
covered by exceptions.rs.
