+++ baml_language/crates/baml_tests/tests/exceptions.rs
@@ -10,7 +10,7 @@
 //!   - Panics vs user throws
 //!   - Nested, rethrow, sequential
 
-use baml_tests::baml_test;
+use baml_tests::{baml_test, baml_test_optimized};
 use bex_engine::BexExternalValue;
 
 /// Assert that execution failed with an uncaught panic of the given class.
@@ -3553,4 +3553,637 @@ async fn caught_panic_has_accessible_fields() {
     }
     ");
     assert_eq!(output.result, Ok(BexExternalValue::Int(7)));
+}
+
+// ============================================================================
+// §11 — Optimized catch dispatch (OptLevel::Two bytecode snapshots)
+// ============================================================================
+
+#[tokio::test]
+async fn catch_four_primitive_type_arms() {
+    // 4 primitive type arms: string, int, bool, null + wildcard.
+    let output = baml_test_optimized!(
+        r#"
+        function fails() -> int {
+            throw "oops"
+        }
+
+        function main() -> int {
+            fails() catch (e) {
+                _: string => 1,
+                _: int => 2,
+                _: bool => 3,
+                _: null => 4,
+                _ => 0
+            }
+        }
+    "#
+    );
+
+    insta::assert_snapshot!(output.bytecode, @r#"
+    function fails() -> int {
+        load_const "oops"
+        throw
+    }
+
+    function main() -> int {
+        call user.fails
+        jump L8
+        load_var e
+        type_tag
+        load_const 1
+        cmp_op ==
+        pop_jump_if_false L0
+        jump L7
+
+      L0:
+        load_var e
+        type_tag
+        load_const 0
+        cmp_op ==
+        pop_jump_if_false L1
+        jump L6
+
+      L1:
+        load_var e
+        type_tag
+        load_const 2
+        cmp_op ==
+        pop_jump_if_false L2
+        jump L5
+
+      L2:
+        load_var e
+        type_tag
+        load_const 3
+        cmp_op ==
+        pop_jump_if_false L3
+        jump L4
+
+      L3:
+        load_var e
+        throw_if_panic
+        load_const 0
+        jump L8
+
+      L4:
+        load_const 4
+        jump L8
+
+      L5:
+        load_const 3
+        jump L8
+
+      L6:
+        load_const 2
+        jump L8
+
+      L7:
+        load_const 1
+
+      L8:
+        return
+    }
+    "#);
+
+    assert_eq!(output.result, Ok(BexExternalValue::Int(1)));
+}
+
+#[tokio::test]
+async fn catch_four_primitive_wildcard_on_float() {
+    // Throw a float — no arm matches, falls through to wildcard.
+    let output = baml_test_optimized!(
+        r#"
+        function fails() -> int {
+            throw 3.14
+        }
+
+        function main() -> int {
+            fails() catch (e) {
+                _: string => 1,
+                _: int => 2,
+                _: bool => 3,
+                _: null => 4,
+                _ => 0
+            }
+        }
+    "#
+    );
+
+    insta::assert_snapshot!(output.bytecode, @"
+    function fails() -> int {
+        load_const 3.14
+        throw
+    }
+
+    function main() -> int {
+        call user.fails
+        jump L8
+        load_var e
+        type_tag
+        load_const 1
+        cmp_op ==
+        pop_jump_if_false L0
+        jump L7
+
+      L0:
+        load_var e
+        type_tag
+        load_const 0
+        cmp_op ==
+        pop_jump_if_false L1
+        jump L6
+
+      L1:
+        load_var e
+        type_tag
+        load_const 2
+        cmp_op ==
+        pop_jump_if_false L2
+        jump L5
+
+      L2:
+        load_var e
+        type_tag
+        load_const 3
+        cmp_op ==
+        pop_jump_if_false L3
+        jump L4
+
+      L3:
+        load_var e
+        throw_if_panic
+        load_const 0
+        jump L8
+
+      L4:
+        load_const 4
+        jump L8
+
+      L5:
+        load_const 3
+        jump L8
+
+      L6:
+        load_const 2
+        jump L8
+
+      L7:
+        load_const 1
+
+      L8:
+        return
+    }
+    ");
+
+    assert_eq!(output.result, Ok(BexExternalValue::Int(0)));
+}
+
+#[tokio::test]
+async fn catch_three_type_arms_bool_throw() {
+    // 3 type arms + wildcard, throw bool.
+    let output = baml_test_optimized!(
+        r#"
+        function fails() -> int {
+            throw true
+        }
+
+        function main() -> int {
+            fails() catch (e) {
+                _: string => 1,
+                _: int => 2,
+                _: bool => 3,
+                _ => 0
+            }
+        }
+    "#
+    );
+
+    insta::assert_snapshot!(output.bytecode, @"
+    function fails() -> int {
+        load_const true
+        throw
+    }
+
+    function main() -> int {
+        call user.fails
+        jump L6
+        load_var e
+        type_tag
+        load_const 1
+        cmp_op ==
+        pop_jump_if_false L0
+        jump L5
+
+      L0:
+        load_var e
+        type_tag
+        load_const 0
+        cmp_op ==
+        pop_jump_if_false L1
+        jump L4
+
+      L1:
+        load_var e
+        type_tag
+        load_const 2
+        cmp_op ==
+        pop_jump_if_false L2
+        jump L3
+
+      L2:
+        load_var e
+        throw_if_panic
+        load_const 0
+        jump L6
+
+      L3:
+        load_const 3
+        jump L6
+
+      L4:
+        load_const 2
+        jump L6
+
+      L5:
+        load_const 1
+
+      L6:
+        return
+    }
+    ");
+
+    assert_eq!(output.result, Ok(BexExternalValue::Int(3)));
+}
+
+#[tokio::test]
+async fn catch_four_user_classes_instanceof_chain() {
+    // User class arms use instanceof checks, not type-tag dispatch.
+    // Even with 4+ arms, these remain sequential instanceof chains.
+    let output = baml_test_optimized!(
+        r#"
+        class AuthError { reason: string }
+        class NotFound { path: string }
+        class RateLimit { retryAfter: int }
+        class Timeout { ms: int }
+
+        function api(mode: int) -> int {
+            if (mode == 0) { throw AuthError { reason: "expired" } }
+            if (mode == 1) { throw NotFound { path: "/users" } }
+            if (mode == 2) { throw RateLimit { retryAfter: 30 } }
+            if (mode == 3) { throw Timeout { ms: 5000 } }
+            throw "unknown"
+        }
+
+        function main() -> int {
+            api(2) catch (e) {
+                _: AuthError => 1,
+                _: NotFound => 2,
+                _: RateLimit => 3,
+                _: Timeout => 4,
+                _ => 0
+            }
+        }
+    "#
+    );
+
+    insta::assert_snapshot!(output.bytecode, @r#"
+    function api(mode: int) -> int {
+        load_var mode
+        load_const 0
+        cmp_op ==
+        pop_jump_if_false L0
+        jump L7
+
+      L0:
+        load_var mode
+        load_const 1
+        cmp_op ==
+        pop_jump_if_false L1
+        jump L6
+
+      L1:
+        load_var mode
+        load_const 2
+        cmp_op ==
+        pop_jump_if_false L2
+        jump L5
+
+      L2:
+        load_var mode
+        load_const 3
+        cmp_op ==
+        pop_jump_if_false L3
+        jump L4
+
+      L3:
+        load_const "unknown"
+        throw
+
+      L4:
+        alloc_instance Timeout
+        load_const 5000
+        init_field .ms
+        throw
+
+      L5:
+        alloc_instance RateLimit
+        load_const 30
+        init_field .retryAfter
+        throw
+
+      L6:
+        alloc_instance NotFound
+        load_const "/users"
+        init_field .path
+        throw
+
+      L7:
+        alloc_instance AuthError
+        load_const "expired"
+        init_field .reason
+        throw
+    }
+
+    function main() -> int {
+        load_const 2
+        call user.api
+        jump L8
+        load_var e
+        load_const AuthError
+        cmp_op instanceof
+        pop_jump_if_false L0
+        jump L7
+
+      L0:
+        load_var e
+        load_const NotFound
+        cmp_op instanceof
+        pop_jump_if_false L1
+        jump L6
+
+      L1:
+        load_var e
+        load_const RateLimit
+        cmp_op instanceof
+        pop_jump_if_false L2
+        jump L5
+
+      L2:
+        load_var e
+        load_const Timeout
+        cmp_op instanceof
+        pop_jump_if_false L3
+        jump L4
+
+      L3:
+        load_var e
+        throw_if_panic
+        load_const 0
+        jump L8
+
+      L4:
+        load_const 4
+        jump L8
+
+      L5:
+        load_const 3
+        jump L8
+
+      L6:
+        load_const 2
+        jump L8
+
+      L7:
+        load_const 1
+
+      L8:
+        return
+    }
+    "#);
+
+    assert_eq!(output.result, Ok(BexExternalValue::Int(3)));
+}
+
+#[tokio::test]
+async fn catch_four_literal_arms() {
+    // Literal patterns use sequential dispatch even with 4+ arms.
+    let output = baml_test_optimized!(
+        r#"
+        function fails(x: int) -> int {
+            if (x == 0) { throw "boom" }
+            if (x == 1) { throw "bang" }
+            if (x == 2) { throw "crash" }
+            throw "other"
+        }
+
+        function main() -> int {
+            fails(1) catch (e) {
+                "boom" => 10,
+                "bang" => 20,
+                "crash" => 30,
+                "other" => 40,
+                _ => 0
+            }
+        }
+    "#
+    );
+
+    insta::assert_snapshot!(output.bytecode, @r#"
+    function fails(x: int) -> int {
+        load_var x
+        load_const 0
+        cmp_op ==
+        pop_jump_if_false L0
+        jump L5
+
+      L0:
+        load_var x
+        load_const 1
+        cmp_op ==
+        pop_jump_if_false L1
+        jump L4
+
+      L1:
+        load_var x
+        load_const 2
+        cmp_op ==
+        pop_jump_if_false L2
+        jump L3
+
+      L2:
+        load_const "other"
+        throw
+
+      L3:
+        load_const "crash"
+        throw
+
+      L4:
+        load_const "bang"
+        throw
+
+      L5:
+        load_const "boom"
+        throw
+    }
+
+    function main() -> int {
+        load_const 1
+        call user.fails
+        jump L8
+        load_var e
+        load_const "boom"
+        cmp_op ==
+        pop_jump_if_false L0
+        jump L7
+
+      L0:
+        load_var e
+        load_const "bang"
+        cmp_op ==
+        pop_jump_if_false L1
+        jump L6
+
+      L1:
+        load_var e
+        load_const "crash"
+        cmp_op ==
+        pop_jump_if_false L2
+        jump L5
+
+      L2:
+        load_var e
+        load_const "other"
+        cmp_op ==
+        pop_jump_if_false L3
+        jump L4
+
+      L3:
+        load_var e
+        throw_if_panic
+        load_const 0
+        jump L8
+
+      L4:
+        load_const 40
+        jump L8
+
+      L5:
+        load_const 30
+        jump L8
+
+      L6:
+        load_const 20
+        jump L8
+
+      L7:
+        load_const 10
+
+      L8:
+        return
+    }
+    "#);
+
+    assert_eq!(output.result, Ok(BexExternalValue::Int(20)));
+}
+
+#[tokio::test]
+async fn catch_mixed_named_and_anonymous_bindings() {
+    // Mixed named (err: ClassName) and anonymous (_: ClassName) bindings
+    // in the same catch block with field access.
+    let output = baml_test_optimized!(
+        r#"
+        class NetworkError { url: string }
+        class AuthError { reason: string }
+        class NotFound { path: string }
+        class RateLimit { retryAfter: int }
+
+        function fails() -> string {
+            throw NetworkError { url: "http://example.com" }
+        }
+
+        function main() -> string {
+            fails() catch (e) {
+                err: NetworkError => err.url,
+                err: AuthError => err.reason,
+                err: NotFound => err.path,
+                _: RateLimit => "rate limited",
+                _ => "unknown"
+            }
+        }
+    "#
+    );
+
+    insta::assert_snapshot!(output.bytecode, @r#"
+    function fails() -> string {
+        alloc_instance NetworkError
+        load_const "http://example.com"
+        init_field .url
+        throw
+    }
+
+    function main() -> string {
+        call user.fails
+        jump L8
+        load_var e
+        load_const NetworkError
+        cmp_op instanceof
+        pop_jump_if_false L0
+        jump L7
+
+      L0:
+        load_var e
+        load_const AuthError
+        cmp_op instanceof
+        pop_jump_if_false L1
+        jump L6
+
+      L1:
+        load_var e
+        load_const NotFound
+        cmp_op instanceof
+        pop_jump_if_false L2
+        jump L5
+
+      L2:
+        load_var e
+        load_const RateLimit
+        cmp_op instanceof
+        pop_jump_if_false L3
+        jump L4
+
+      L3:
+        load_var e
+        throw_if_panic
+        load_const "unknown"
+        jump L8
+
+      L4:
+        load_const "rate limited"
+        jump L8
+
+      L5:
+        load_var e
+        load_field .path
+        jump L8
+
+      L6:
+        load_var e
+        load_field .reason
+        jump L8
+
+      L7:
+        load_var e
+        load_field .url
+
+      L8:
+        return
+    }
+    "#);
+
+    assert_eq!(
+        output.result,
+        Ok(BexExternalValue::String("http://example.com".into()))
+    );
 }
